feat: extract form SignUpSchema to own file
